replaced hideous instructions subform with nicer one
